+++ test_common/lib/test_sdk_configuration.dart
@@ -84,7 +84,9 @@ class TestSdkConfigurationProvider extends SdkConfigurationProvider {
       }
     } catch (e, s) {
       _logger.warning('Failed delete SDK directory copy', e, s);
-      dispose(retry: false);
+      if (retry) {
+        dispose(retry: false);
+      }
     }
   }
 }