+++ test_common/lib/test_sdk_layout.dart
@@ -131,11 +131,16 @@ class TestSdkLayout {
           'bin',
           Platform.isWindows ? 'dart.exe' : 'dart',
         ),
+        dartAotRuntimePath: p.join(
+          sdkLayout.sdkDirectory,
+          'bin',
+          Platform.isWindows ? 'dartaotruntime.exe' : 'dartaotruntime',
+        ),
         frontendServerSnapshotPath: p.join(
           sdkLayout.sdkDirectory,
           'bin',
           'snapshots',
-          'frontend_server.dart.snapshot',
+          'frontend_server_aot.dart.snapshot',
         ),
         dartdevcSnapshotPath: sdkLayout.dartdevcSnapshotPath,
         kernelWorkerSnapshotPath: p.join(
@@ -187,6 +192,7 @@ class TestSdkLayout {
   final String stackTraceMapperPath;
 
   final String dartPath;
+  final String dartAotRuntimePath;
   final String frontendServerSnapshotPath;
   final String dartdevcSnapshotPath;
   final String kernelWorkerSnapshotPath;
@@ -210,6 +216,7 @@ class TestSdkLayout {
     required this.requireJsPath,
     required this.stackTraceMapperPath,
     required this.dartPath,
+    required this.dartAotRuntimePath,
     required this.frontendServerSnapshotPath,
     required this.dartdevcSnapshotPath,
     required this.kernelWorkerSnapshotPath,